Have reader remember index and offset of each match
Additionally, this changes the invariants of reader_history_t::matches_:
it can now be empty, and is being used as a proper double-ended queue. It
already was a std::deque but only appended to. Now we prepend and append
search results from forward and backward searches respectively.

Change readline command beginning-of-history to move to the oldest search
match. This seems like a alright fix because I think it used to do the
same thing as end-of-history (move back to the original search string).