Expose history search's current item index
The reader caches history search matches.  However, when we allow editing
the search string, an edit invalidates all matches.
Since the history search's current item may be different from the reader's
current item (because we allow moving around the cached matches), the
reader needs to be able to reset the history search's index to the one
that corresponds to the reader's item.