Make reader history match_index a maybe_t
When the history search has no match, we used match index zero.
The first item in the list of cached matches would always be the
search string, so that worked nicely.
However, including the search string ther seems redundant
(we'd need to duplicate edits), so we'll move away from that.
The first step is to make 0 a valid match index.